requirements.txt
requests==2.21.0
RPi.GPIO==0.7.0
six==1.12.0
urllib3==1.24.1
schedule==0.6.0
pystache==0.5.4
PyYAML>=5.1
